improved styling and added redirects from login page
- changed button styling
- clicking on "sign up" now redirects to loginSelect page
- each button redirects to a different page
